requirements.txt
# Flask and extensions
Flask==3.0.3
Flask-Cors==4.0.0
python-dotenv==1.0.1

# MySQL
mysql-connector-python==8.3.0

# Machine Learning
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2

# GenAI - Choose one based on what you use
openai==1.34.0       # For OpenAI Chatbot
google-generativeai==0.5.4  # For Gemini Chatbot

# Security / Utils (optional)
bcrypt==4.1.2